+++ esibd/devices/lakeshore335/lakeshore.py
@@ -69,6 +69,7 @@ class LS335(Device):
 
     def changeUnit(self) -> None:
         """Update plots to account for change of unit."""
+        self.pluginManager.connectAllSources()
         if self.liveDisplayActive():
             self.clearPlot()
             self.liveDisplay.plot()